Restructered according to -Algorithm Type, Algorith Name, Language, Source Code-
