+++ Circle Drawing/Bresenham/C++/circle.cc
