+++ Sorting/Quick Sort/C++/quick_sort.cpp
