+++ Sorting/Stooge Sort/C++/stooge_sort.cpp
