+++ Stable Matching/Gale Shapley/Javascript/galeshapley.js
