+++ String Matching/Rabin Karp/Javascript/RabinKarp.js
